+++ core/server/api_container/server/api_container_service.go
@@ -54,6 +54,10 @@ const (
 	// Overwrite existing module with new module, this allows user to iterate on an enclave with a
 	// given module
 	doOverwriteExistingModule = true
+
+	isScript    = true
+	isNotScript = false
+	isNotRemote = false
 )
 
 // Guaranteed (by a unit test) to be a 1:1 mapping between API port protos and port spec protos
@@ -183,6 +187,12 @@ func (apicService ApiContainerService) RunStarlarkScript(args *kurtosis_core_rpc
 	serializedStarlarkScript := args.GetSerializedScript()
 	serializedParams := args.GetSerializedParams()
 	dryRun := shared_utils.GetOrDefaultBool(args.DryRun, defaultStartosisDryRun)
+
+	if err := apicService.metricsClient.TrackKurtosisRun(startosis_constants.PackageIdPlaceholderForStandaloneScript, isNotRemote, dryRun, isScript); err != nil {
+		//We don't want to interrupt users flow if something fails when tracking metrics
+		logrus.Errorf("An error occurred tracking kurtosis run event\n%v", err)
+	}
+
 	apicService.runStarlark(dryRun, startosis_constants.PackageIdPlaceholderForStandaloneScript, serializedStarlarkScript, serializedParams, stream)
 	return nil
 }
@@ -193,6 +203,11 @@ func (apicService ApiContainerService) RunStarlarkPackage(args *kurtosis_core_rp
 	moduleContentIfLocal := args.GetLocal()
 	serializedParams := args.SerializedParams
 	dryRun := shared_utils.GetOrDefaultBool(args.DryRun, defaultStartosisDryRun)
+
+	if err := apicService.metricsClient.TrackKurtosisRun(packageId, isRemote, dryRun, isNotScript); err != nil {
+		//We don't want to interrupt users flow if something fails when tracking metrics
+		logrus.Errorf("An error occurred tracking kurtosis run event\n%v", err)
+	}
 
 	scriptWithRunFunction, interpretationError := apicService.runStarlarkPackageSetup(packageId, isRemote, moduleContentIfLocal)
 	if interpretationError != nil {